pyproject.toml
[tool.poetry]
name = "mpscraper"
version = "0.2.3"
description = "Marktplaats.nl (Dutch Classifieds) Listing Scraper"
authors = ["Ross <git@ross.ch>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
mpscraper = "mpscraper.__main__:main"
mpbot = "telegram_bot.client_bot:run_client_bot"
mpclient = "telegram_bot.client_bot:run_client_bot"
mpadmin = "telegram_bot.admin_bot:run_admin_bot"

[tool.poetry.dependencies]
python = "^3.13"
beautifulsoup4 = "^4.14.2"
selenium = "^4.35.0"
PyVirtualDisplay = "^3.0"
tqdm = "^4.67.1"
pandas = "^2.3.2"
lxml = "^6.0.2"
requests = "^2.32.0"
aiogram = "^3.15"
python-dotenv = "^1.0"
fake-http-header = "^0.3.5"

[tool.poetry.group.dev.dependencies]
ruff = "^0.13.2"
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
bandit = "^1.8.6"
pre-commit = "^4.3.0"

[tool.ruff]
line-length = 100

[tool.bandit]
exclude = ["test_*.py"]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
